+++ src/org/uecide/gui/swing/SketchHeaderFilesNode.java
@@ -13,6 +13,7 @@ import javax.swing.JTree;
 import java.io.IOException;
 
 import java.awt.Font;
+import java.awt.Component;
 
 public class SketchHeaderFilesNode extends SketchTreeNodeBase {
     public SketchHeaderFilesNode(Context c, SketchTreeModel m) {
@@ -48,8 +49,13 @@ public class SketchHeaderFilesNode extends SketchTreeNodeBase {
     public void performDoubleClick() {
     }
 
-    public Font getFont() {
-        return new JLabel().getFont();
+    public Component getRenderComponent(JLabel original, JTree tree) {
+        try {
+            original.setIcon(getIcon(tree));
+        } catch (Exception ex) {
+        }
+
+        return original;
     }
 
 }